renamed verify_permissions -> permissions
