+++ rust/holo-public-api/src/providers/auth/mod.rs
@@ -3,10 +3,10 @@ mod get_refresh_token_version;
 mod hash_apikey;
 mod roles;
 mod sign_tokens;
-mod verify_permissions;
+mod permissions;
 
 pub use common::{generate_api_key, get_apikey_from_headers};
 pub use get_refresh_token_version::get_refresh_token_version;
 pub use hash_apikey::hash_apikey;
 pub use sign_tokens::{sign_tokens, SignJwtTokenOptions};
-pub use verify_permissions::{get_all_accessible_owners_from_permissions, verify_all_permissions};
+pub use permissions::{get_all_accessible_owners_from_permissions, verify_all_permissions};
